refactor(react): remove circular references
