+++ packages/react/src/ActionList/Group.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import {useId} from '../hooks/useId'
 import Box from '../Box'
 import {SxProp, BetterSystemStyleObject, merge} from '../sx'
-import {ListContext, ActionListProps} from './List'
+import {ListContext, type ActionListProps} from './shared'
 import {AriaRole} from '../utils/types'
 import {default as Heading} from '../Heading'
 import type {ActionListHeadingProps} from './Heading'
