+++ packages/react/src/ActionList/index.ts
@@ -7,7 +7,7 @@ import {Description} from './Description'
 import {LeadingVisual, TrailingVisual} from './Visuals'
 import {Heading} from './Heading'
 
-export type {ActionListProps} from './List'
+export type {ActionListProps} from './shared'
 export type {ActionListGroupProps} from './Group'
 export type {ActionListItemProps} from './shared'
 export type {ActionListLinkItemProps} from './LinkItem'
